requirements.txt
Django~=2.1.7
mysqlclient~=1.4.2.post1
pytz~=2018.9
wheel~=0.30.0
paho-mqtt~=1.1
pyjwt~=1.7.1
google-cloud-pubsub~=0.39
cryptography==2.5
django-octicons~=1.0.2
django-tables2~=2.0.6
django-tempus-dominus~=5.1.2.5

